Tweak layout
I loved the "pop out" effect of the graphs being wider than the other
text, but a lot of people seem to be confused and think it's a bug,
so now it's flush.

On the upside, this involved making the padding of the sections super
wide, which let us unindent the headings for a pretty cool effect.

I also compressed the header, because everyone's very proud of their
text and wants more room for it.

I also made a JPEG of the header image. 600 KB instead of 6 MB. Still
huge but not, like, the entire rest of the site combined. I like it
huge, it's very detailed on Retina screens.